requirements.txt
# Core ADK dependencies
google-adk>=0.1.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Streamlit for the web interface
streamlit>=1.28.0

# HTTP requests
requests>=2.31.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Timezone handling
pytz>=2023.3

# Environment variables
python-dotenv>=1.0.0

# CORS support
fastapi-cors>=0.0.6 